documind/backend/requirements.txt
# ./backend/requirements.txt

fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
asyncpg
boto3

# PDF Processing
pdfplumber

# AI / Vector DB
langchain
langchain-community
langchain-huggingface
langchain-pinecone
sentence-transformers
pinecone

# Local LLM libraries
torch
transformers
accelerate

# Utilities
python-dotenv
python-multipart
pydantic[email]
minio
langchain
sentence-transformers
pinecone-client
redis
openai
tiktoken
langchain-openai
langchain-pinecone